+++ tests/unit/tests/Form/LegacyConsumer/Commands/SetupFieldPersistenceTest.php
@@ -2,9 +2,9 @@
 
 use Give\Framework\FieldsAPI\Text;
 
-final class SetupFieldPersistanceTest extends Give_Unit_Test_Case {
+final class SetupFieldPersistenceTest extends Give_Unit_Test_Case {
 
-    public function testFieldPersistance() {
+    public function testFieldPersistence() {
 
         add_action( 'give_fields_donation_form', function( $form ) {
             $form->append(